Optimizer changes + Rprop.

+++ src/nl/rug/joptimize/opt/optimizers/SGD.java
@@ -32,7 +32,7 @@ public class SGD<ParamType extends OptParam<ParamType>> extends
     protected ParamType sgdEpoch(SeparableCostFunction<ParamType> cf, ParamType params, double learningRate) {
         ParamType out = params.copy();
         int size = cf.size();
-        effectiveLearningRate = learningRate / (1+ t/tMax);
+        effectiveLearningRate = learningRate / (1+ t/(double)tMax);
         
         for (int i = 0; i < size; i++) {
             out.sub_s(cf.deriv(out, rand.nextInt(size)).multiply_s(effectiveLearningRate));
